requirements.txt
Django==3.2
tweepy==3.10.0
django-redis==4.12.1